password validator component
